# internal/http/openapi.yaml
openapi: 3.0.3
info:
  title: chino-api
  version: 0.1.0
  description: Backend API for the chino product of the zaentrum platform.
servers:
  - url: /api
paths:
...
  /v1/me:
    get:
      summary: Echo the caller's OIDC subject
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /v1/items:
    get:
      summary: List items for this product (stub during beta)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /v1/items/{id}:
    get:
      summary: An item with its genres, cast, subtitles, trailers and segments
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
        "404":
          description: No such item
        "502":
          description: The catalog failed
  /v1/people:
    get:
      summary: Search people by name
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: empty answers an empty list
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  people:
                    type: array
                    items: { $ref: "#/components/schemas/Person" }
                  total: { type: integer }
        "502":
          description: The catalog failed
  /v1/people/{id}:
    get:
      summary: A person, their details and their filmography
      description: >
        The filmography is one card per title, newest first, each with the
        person's roles on it. The biography is in the language `lang` names,
        else the first of Accept-Language's the catalog has, else English,
        else any; the response varies on Accept-Language.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
        - in: query
          name: lang
          schema: { type: string }
          description: biography language (a language tag; its primary subtag counts)
        - in: header
          name: Accept-Language
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PersonDetail" }
        "404":
          description: No such person
        "502":
          description: The catalog failed
  /v1/people/{id}/profile:
    get:
      summary: A person's portrait (what profile_url points at)
      description: >
        Proxied from the catalog manager with the posters' auth: a bearer
        or ?stream= (a stream token), so an <img src> works (?token=, the
        bearer in the URL, is deprecated and goes in the next release). Range and
        If-None-Match pass through; so do the upstream's status and headers
        (ETag, 304, 404).
      security: [{ bearerAuth: [] }, { streamToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The image
          content:
            image/*:
              schema: { type: string, format: binary }
        "304":
          description: Not modified (If-None-Match matched the ETag)
        "404":
          description: The person has no portrait
  /v1/me/watchlists:
    get:
      summary: List the caller's named watchlists
      description: Default list first, then others by createdAt ascending.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  lists:
                    type: array
                    items: { $ref: "#/components/schemas/Watchlist" }
        "401":
          description: Unauthorized
    post:
      summary: Create a new named watchlist
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 60 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Watchlist" }
        "400":
          description: Empty / over-length name
        "409":
          description: Too many lists (max 50) or case-insensitive duplicate name
  /v1/me/watchlists/memberships:
    get:
      summary: Which of the caller's lists each item belongs to
      description: >
        Powers the add-to-list picker checkmarks and the card "saved" badge.
        Items not in any list may be omitted or map to an empty array.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ids
          required: true
          schema: { type: string }
          description: comma-separated item ids (max ~100)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberships:
                    type: object
                    additionalProperties:
                      type: array
                      items: { type: string }
        "401":
          description: Unauthorized
  /v1/me/watchlists/{listId}:
    get:
      summary: A single list's metadata + item ids
      description: Items newest-added first.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: listId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string }
                  isDefault: { type: boolean }
                  items:
                    type: array
                    items: { type: string }
        "404":
          description: Not the caller's list
    patch:
      summary: Rename a list
      description: Renaming the default list keeps isDefault=true.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: listId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 60 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Watchlist" }
        "400":
          description: Empty / over-length name
        "404":
          description: Not the caller's list
        "409":
          description: Case-insensitive duplicate name
    delete:
      summary: Delete a list (cascades its items)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: listId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not the caller's list
        "409":
          description: Cannot delete the default list
  /v1/me/watchlists/{listId}/items/{itemId}:
    put:
      summary: Add an item to a list (idempotent)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: listId
          required: true
          schema: { type: string }
        - in: path
          name: itemId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Added (or already present)
        "404":
          description: Not the caller's list
    delete:
      summary: Remove an item from a list (idempotent)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: listId
          required: true
          schema: { type: string }
        - in: path
          name: itemId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed (or already absent)
        "404":
          description: Not the caller's list
  /v1/me/watchlist:
    get:
      summary: The default list's item ids (back-compat alias)
      description: >
        Back-compat alias kept for already-installed mobile/TV builds. Now
        maps to the caller's default list (lazily created on first access).
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { type: string }
        "401":
          description: Unauthorized
  /v1/me/watchlist/{id}:
    put:
      summary: Add an item to the default list (back-compat alias)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Added (or already present)
    delete:
      summary: Remove an item from the default list (back-compat alias)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed (or already absent)
...
components:
  schemas:
    Item:
      type: object
      description: The fields a detail page uses; others are omitted here.
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        year: { type: integer }
        poster_url: { type: string }
        backdrop_url: { type: string }
        genres:
          type: array
          items: { type: string }
        cast:
          description: >
            As the catalog orders and caps it: role by role (actor, creator,
            director, writer, producer, composer, cinematographer, editor,
            then any other role alphabetically), billing order within a role;
            at most 20 actors and 10 people of every other role.
          type: array
          items: { $ref: "#/components/schemas/CastEntry" }
        roles:
          description: on a person's filmography only, the person's roles on the item
          type: array
          items: { type: string }
        watched_at: { type: string, format: date-time }
...
    Person:
      type: object
      required: [id, name, has_profile]
      properties:
        id: { type: string }
        name: { type: string }
        credits: { type: integer, description: "search results: the number of titles" }
        has_profile: { type: boolean }
        profile_url:
          type: string
          description: set when has_profile; /api/v1/people/{id}/profile
    PersonDetail:
      allOf:
        - $ref: "#/components/schemas/Person"
        - type: object
          description: Every optional field is omitted when unknown.
          properties:
            sort_name: { type: string }
            also_known_as:
              type: array
              items: { type: string }
            birth_date: { type: string, format: date }
            death_date: { type: string, format: date }
            birthplace: { type: string }
            known_for_department: { type: string }
            biography: { type: string }
            biography_lang: { type: string, description: "a primary language subtag" }
            tmdb_person_id: { type: string }
            imdb_id: { type: string }
            items:
              type: array
              items: { $ref: "#/components/schemas/Item" }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    streamToken:
      type: apiKey
      in: query
      name: stream
      description: a stream token from POST /api/v1/me/stream-token